requirements.txt
gspread
google-auth
python-dotenv
pandas
requests
